refactor: 💡 Moved files, removed unused
