+++ lib/App.jsx
@@ -4,17 +4,17 @@ import IntroView from './IntroView'
 import UserView from './UserView'
 import ReportView from './ReportView'
 import PageTracker from './components/PageTracker'
-import { getText } from './form/appTexts'
+import { getText } from './appTexts'
 
 export default () => {
   return (
     <Router>
       <nav className="navbar navbar-light bg-light">
         <a href="https://seb.se/">
-          <div className="navbar-brand">{getText('label_navleft')}</div>
+          <div className="navbar-brand">{getText('label_navprimary')}</div>
         </a>
         <a href="/">
-          <span className="navbar-text float-left">{getText('label_navright')}</span>
+          <span className="navbar-text float-left">{getText('label_navsecondary')}</span>
         </a>
       </nav>
       <PageTracker>
